Fixed #2458 - Retention Index to Time - allow float for RI

+++ chemclipse/plugins/org.eclipse.chemclipse.model/src/org/eclipse/chemclipse/model/support/RetentionIndexMap.java
@@ -91,7 +91,7 @@ public class RetentionIndexMap {
 	 * @param retentionIndex
 	 * @return int
 	 */
-	public int getRetentionTime(int retentionIndex) {
+	public int getRetentionTime(float retentionIndex) {
 
 		return RetentionIndexMath.calculateRetentionTime(retentionIndex, retentionIndexMap);
 	}
